@@ -17,9 +17,9 @@
     "bootstrap": "^5.3.3",
     "classnames": "^2.3.1",
     "contentful": "^10.13.2",
-    "react": "^17.0.0",
+    "react": "^18.3.1",
     "react-bootstrap": "^2.10.4",
-    "react-dom": "^17.0.0",
+    "react-dom": "^18.3.1",
     "react-router-dom": "^6.26.0"
   },
   "devDependencies": {
